pyproject.toml
[project]
name = "gmail-mcp-secure"
version = "0.1.0"
description = "A secure, minimal MCP server providing Gmail access to AI assistants"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "stackdev-opensource"},
]
keywords = ["mcp", "gmail", "ai", "claude", "model-context-protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Email",
]
dependencies = [
    "mcp>=1.3.0",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.154.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.0",
    "commitizen>=4.0.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
gmail-mcp = "gmail_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/stackdev-opensource/gmail-mcp"
Repository = "https://github.com/stackdev-opensource/gmail-mcp"
Issues = "https://github.com/stackdev-opensource/gmail-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gmail_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "SIM"]
ignore = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]

# -- Commitizen: conventional commits + auto-versioning --
[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = [
    "pyproject.toml:^version",
    "src/gmail_mcp/__init__.py:^__version__",
]
tag_format = "v$version"
update_changelog_on_bump = true
changelog_file = "CHANGELOG.md"
